+++ bin/calculate_tmb.py
@@ -3,6 +3,7 @@
 
 import os
 import argparse
+import pyranges
 
 from utils import *
 
@@ -63,6 +64,30 @@ def _parse_args():
 
     args = parser.parse_args()
     return args
+
+
+def calculate_nmd(out, gr_nmd, filter_variant_type, filter_variant_classification):
+    if filter_variant_classification:
+        maf_gr = out[
+            (out["Variant_Type"].isin(filter_variant_type))
+            & (out["Variant_Classification"].isin(filter_variant_classification))
+        ][["Chromosome", "Start_Position", "End_Position"]]
+    else:
+        maf_gr = out[(out["Variant_Type"].isin(filter_variant_type))][
+            ["Chromosome", "Start_Position", "End_Position"]
+        ]
+    if len(maf_gr):
+        maf_gr.columns = ["Chromosome", "Start", "End"]
+        maf_gr = pyranges.PyRanges(df=maf_gr)
+        tmb_nmd_escapees = np.sum(
+            pyranges.count_overlaps(
+                grs={"nmd": gr_nmd}, how="containment", features=maf_gr
+            ).nmd
+            != 0
+        )
+        return tmb_nmd_escapees
+
+    return 0.0
 
 
 def main():
